Move resource.h to MUMain/Resources/Windows subdirectory
