+++ Source Main 5.2/source/MUMain/Resources/Windows/resource.h
