+++ Source Main 5.2/source/MUMain/Winmain.cpp
@@ -23,7 +23,7 @@
 #include "ZzzLodTerrain.h"
 #include "DSPlaySound.h"
 
-#include "resource.h"
+#include "MUMain/Resources/Windows/resource.h"
 #include <imm.h>
 #include "ZzzPath.h"
 #include "Local.h"
