+++ Source Main 5.2/source/resource.rc
@@ -1,6 +1,6 @@
 //Microsoft Developer Studio generated resource script.
 //
-#include "resource.h"
+#include "MUMain/Resources/Windows/resource.h"
 
 #define APSTUDIO_READONLY_SYMBOLS
 /////////////////////////////////////////////////////////////////////////////
@@ -29,7 +29,7 @@ LANGUAGE LANG_KOREAN, SUBLANG_DEFAULT
 
 1 TEXTINCLUDE MOVEABLE PURE
 BEGIN
-"resource.h\0"
+"MUMain/Resources/Windows/resource.h\0"
 END
 
 2 TEXTINCLUDE MOVEABLE PURE
@@ -74,7 +74,7 @@ VALUE "CompanyName", L"WebZen\0"
 VALUE "FileDescription", L"main\0"
 VALUE "FileVersion", L"1, 3, 34, 0\0"
 VALUE "InternalName", L"main\0"
-VALUE "LegalCopyright", L"Copyright ¨Ï 2002\0"
+VALUE "LegalCopyright", L"Copyright ï¿½ï¿½ 2002\0"
 VALUE "LegalTrademarks", L"\0"
 VALUE "OriginalFilename", L"main.exe\0"
 VALUE "PrivateBuild", L"\0"
